delete components from pages module
