--- apps/modernization-ui/src/pages/advancedSearch/AdvancedSearch.scss
@@ -1,196 +1,0 @@
-@use 'styles/colors';
-
-button {
-    white-space: nowrap;
-    img {
-        margin-left: 6px;
-    }
-}
-.break-workd {
-    word-break: normal !important;
-}
-.search-page-height {
-    position: relative;
-    .accordian-item {
-        padding-top: 0;
-    }
-}
-
-.bg-light {
-    background-color: colors.$background;
-}
-
-.cursor-pointer {
-    cursor: pointer;
-}
-
-.type {
-    border-bottom: 0.4375rem solid transparent;
-}
-.type.active {
-    border-bottom: 0.4375rem solid #005ea2;
-}
-
-.patient-profile-tab {
-    font-size: 1.375rem;
-}
-
-.left-searchbar {
-    position: relative;
-    width: 100%;
-    height: calc(100vh - 152.73px);
-    overflow-y: auto;
-    padding-bottom: 15px;
-}
-
-.confirm-bage {
-    width: fit-content;
-}
-.usa-pagination .usa-current {
-    background-color: #003470 !important;
-}
-
-.refine-text {
-    font-size: 22px;
-    padding: 25px 1rem 0;
-}
-
-.bottom-search {
-    border-top: 1px solid colors.$border;
-}
-
-.clear-btn {
-    margin-top: 9px !important;
-}
-
-.bottom-search {
-    padding-top: 6px;
-    padding-bottom: 15px;
-    background-color: colors.$base-white;
-    position: absolute;
-    bottom: 0;
-    width: 100%;
-}
-
-.scrollable-results {
-    height: calc(100vh - 152.73px);
-    overflow-y: auto;
-}
-
-.advanced-search {
-    .advanced-search-message {
-        a {
-            cursor: pointer;
-            border-bottom: 1px solid;
-        }
-    }
-    .grid-row.grid-gap-6 > * {
-        padding-left: 1.5rem !important;
-        padding-right: 1.5rem !important;
-        width: 22% !important;
-    }
-
-    .grid-row.grid-gap-6 .large-col {
-        width: 32% !important;
-    }
-}
-
-.button-group {
-    position: relative;
-    min-height: 50px;
-
-    .usa-nav__submenu {
-        position: absolute;
-        right: 0.5rem;
-        top: 100%;
-        background-color: colors.$base-white;
-        width: max-content;
-        filter: drop-shadow(0px 4px 14px rgba(0, 0, 0, 0.15));
-        border-radius: 5px;
-        z-index: 1;
-
-        .usa-nav__submenu-item {
-            border-top: none !important;
-            border-bottom: 1px solid #dfdfdf;
-            padding-bottom: 10px;
-            margin: 15px;
-            display: flex;
-            .usa-button--unstyled {
-                margin: 0;
-                color: black;
-                text-decoration: none !important;
-            }
-        }
-        .usa-nav__submenu-item:last-child {
-            border: none;
-            margin-bottom: 5px;
-        }
-    }
-}
-
-.searchWrapper {
-    border: 1px solid #565c65 !important;
-    border-radius: 0 !important;
-    padding-right: 25px;
-}
-
-.multiSelectContainer li {
-    font-size: 14px;
-}
-
-.highlightOption {
-    background: #dee7f5;
-    color: #005ea2;
-}
-
-.chip {
-    background: #dee7f5;
-    color: #005ea2;
-    font-size: 16px;
-    white-space: pre-wrap;
-}
-
-.multiselect {
-    background-image: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIGhlaWdodD0iMjQiIHZpZXdCb3g9IjAgMCAyNCAyNCIgd2lkdGg9IjI0Ij48cGF0aCBkPSJNMCAwaDI0djI0SDB6IiBmaWxsPSJub25lIi8+PHBhdGggZD0iTTEyIDUuODNMMTUuMTcgOWwxLjQxLTEuNDFMMTIgMyA3LjQxIDcuNTkgOC44MyA5IDEyIDUuODN6bTAgMTIuMzRMOC44MyAxNWwtMS40MSAxLjQxTDEyIDIxbDQuNTktNC41OUwxNS4xNyAxNSAxMiAxOC4xN3oiLz48L3N2Zz4=),
-        linear-gradient(transparent, transparent);
-    background-repeat: no-repeat;
-    appearance: none;
-    background-color: #fff;
-    background-position: right 0.5rem center;
-    background-size: 1.25rem;
-}
-
-.multiSelectContainer input {
-    &::placeholder {
-        font-weight: normal !important;
-        color: black !important;
-    }
-    &:focus {
-        outline: none !important;
-    }
-}
-
-.status {
-    color: colors.$base-white;
-    border-radius: 22px;
-    width: max-content;
-    padding: 2px 8px;
-    font-size: 12px;
-}
-
-.fill-height {
-    height: 100%;
-    align-content: space-between;
-}
-.checkbox-input {
-    margin-right: 15px;
-    text-transform: capitalize !important;
-}
-
-.search-result-item-label {
-    font-size: 0.8125rem;
-    font-weight: 600;
-    letter-spacing: 0.02031rem;
-    text-transform: uppercase;
-    color: colors.$base;
-}
