--- apps/modernization-ui/src/pages/advancedSearch/components/chips/Chip.spec.tsx
@@ -1,6 +1,0 @@
-import Chip from './Chip';
-
-describe('Chip component tests', () => {
-    it('should render a chip icon', () => {
-    });
-});
